Add navigation links to header and create RandomImage component for dynamic image fetching
